requirements.txt
asgiref==3.7.2
coverage==7.3.1
Django==4.2.5
django-redis==5.3.0
djangorestframework==3.14.0
mysqlclient==2.2.0
Pillow==10.0.0
pytz==2023.3.post1
redis==5.0.0
sqlparse==0.4.4
